before go home
