+++ philo/philo.h
@@ -11,53 +11,50 @@
 # define INT_MAX			2147483647
 # define INT_MIN			-2147483648
 
-typedef	struct s_arg
+typedef struct s_status
 {
-	int		philo_n;
-	size_t	life_t;
-	size_t	eat_t;
-	size_t	sleep_t;
-	int		eat_cnt;
-}		t_arg;
+	int	end;
+	int	n_full;
+}	t_status;
 
-typedef struct s_flag
+typedef struct s_mutex
 {
-	int				eat_f;
-	int				die_f;
-	int				err_f;
-	int				start_f;
-}					t_flag;
-
-typedef struct s_info
-{
-	t_arg			arg;
-	pthread_mutex_t	*fork_mutex;
-	pthread_mutex_t prt_mutex;
-	pthread_mutex_t	t_mutex;
-	t_flag			flags;
-	size_t			start_time;
-}					t_info;
+	pthread_mutex_t	*fork;
+	pthread_mutex_t print;
+	pthread_mutex_t	sub;
+}	t_mutex;
 
 typedef struct s_philo
 {
-	int			idx;
-	int			p_eat_cnt;
-	size_t		life_time;
-	t_info		*info;
+	int 		idx;
 	pthread_t	tid;
-}				t_philo;
+	int			cnt_eat;
+	size_t		life_time;
+}	t_philo;
 
-enum	e_action
+typedef struct s_arg
+{
+	int	n_philo;
+	int t2die;
+	int t2eat;
+	int t2sleep;
+	int must_eat;
+}	t_arg;
+
+typedef struct s_info
+{
+	t_arg		arg;
+	t_philo 	*philo;
+	t_mutex 	mutex;
+	t_status 	stat;
+	int			at;
+}	t_info;
+
+enum e_enum
 {
 	ERROR = -1,
-	SUCCESS,
-	FORK,
-	EATING,
-	SLEEPING,
-	THINKING,
-	DIED
+	SUCCESS = 1
 };
 
 long long	ft_atoi(const	char	*str);
-
 #endif
